nota_venta_web/backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25
pymysql==1.1.0
pyodbc==5.0.1
alembic==1.13.1

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Procesamiento de archivos
openpyxl==3.1.2
pandas==2.2.0

# Tareas asíncronas
celery==5.3.6
redis==5.0.1

# WebSockets
python-socketio==5.11.0
websockets==12.0

# SMTP
aiosmtplib==3.0.1

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# Utilidades
python-dateutil==2.8.2
